implement AccessTokenInteractor(#17)
