+++ interactor/ranking_interactor.py
@@ -3,11 +3,10 @@ import time
 from injector import inject, singleton
 from logging import Logger
 import requests
-from requests.auth import HTTPBasicAuth
 
 from envs import Envs
+from interface.usecase.access_token_usecase import AccessTokenUsecase
 from interface.usecase.ranking_usecase import RankingUsecase
-from interface.repository.access_token_repository import AccessTokenRepository
 from interface.repository.ranking_repository import RankingRepository
 
 
@@ -17,12 +16,12 @@ class RankingInteractor(RankingUsecase):
     def __init__(
         self,
         env: Envs,
-        access_token_repository: AccessTokenRepository,
+        access_token_useacse: AccessTokenUsecase,
         ranking_repository: RankingRepository,
         logger: Logger,
     ) -> None:
         self.env = env
-        self.access_token_repository = access_token_repository
+        self.access_token_useacse = access_token_useacse
         self.ranking_repository = ranking_repository
         self.logger = logger
 
@@ -44,19 +43,7 @@ class RankingInteractor(RankingUsecase):
         return ranking
 
     def _create_ranking(self) -> list:
-        if self.access_token_repository.exist():
-            access_token_ttl = self.access_token_repository.get_ttl()
-
-            # check if access token is expired
-            if access_token_ttl > time.time():
-                access_token = self.access_token_repository.get()
-            else:
-                self.logger.info("access_token is expired and create new one.")
-                access_token = self._create_spotify_access_token()
-
-        else:
-            self.logger.info("access_token doesn't exist and create new one.")
-            access_token = self._create_spotify_access_token()
+        access_token = self.access_token_usecase.get_access_token()
 
         global_charts_id = "37i9dQZEVXbMDoHDwVN2tF"
         response = requests.get(
@@ -112,21 +99,3 @@ class RankingInteractor(RankingUsecase):
         self.ranking_repository.create(ranking, ttl)
         return ranking
 
-    def _create_spotify_access_token(self) -> str:
-        response = requests.post(
-            "https://accounts.spotify.com/api/token",
-            headers={"Content-Type": "application/x-www-form-urlencoded"},
-            auth=HTTPBasicAuth(self.env.CLIENT_ID, self.env.CLIENT_SECRET),
-            data={"grant_type": "client_credentials"},
-        )
-
-        if response.status_code != requests.codes.ok:
-            self.logger.error(response.json())
-            raise RuntimeError("cannot fetch access_token correctly.")
-
-        access_token = response.json()["access_token"]
-
-        # spotify access token will expire after 1 hour and set ttl to 50 minutes
-        ttl = time.time() + 60 * 50
-        self.access_token_repository.create(access_token, ttl)
-        return access_token
