+++ interface/usecase/access_token_usecase.py
@@ -1,0 +1,7 @@
+from abc import ABCMeta, abstractmethod
+
+
+class AccessTokenUsecase(metaclass=ABCMeta):
+    @abstractmethod
+    def get_access_token(self) -> str:
+        pass
